# storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Identityservice API Swagger JWT Laravel",
        "version": "1.0.0"
    },
    "servers": [
        {
            "url": "/api/"
        }
    ],
    "paths": {
...
        "/products/{product_id}": {
            "get": {
                "tags": [
                    "Products"
                ],
                "summary": "Show a Product from his Id",
                "description": "Show a Product from his Id",
                "operationId": "showProduct",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Product--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Product"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "delete": {
                "tags": [
                    "Products"
                ],
                "summary": "Delete a Product",
                "description": "Delete a Product",
                "operationId": "destroyProduct",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Product--id"
                    }
                ],
                "responses": {
                    "204": {
                        "description": "No content"
                    },
                    "404": {
                        "description": "Product not found"
                    }
                }
            },
            "patch": {
                "tags": [
                    "Products"
                ],
                "summary": "Update an existing Product",
                "description": "Update an existing Product",
                "operationId": "updateProduct",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Product--id"
                    }
                ],
                "requestBody": {
                    "description": "Product to update",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Product"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "204": {
                        "description": "No content"
                    }
                }
            }
        },
        "/roles": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Get list of Roles",
                "description": "Returns list of Role",
                "operationId": "indexRole",
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "$ref": "#/components/schemas/Roles"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Roles"
                ],
                "summary": "Insert a new Role",
                "description": "Insert a new Role",
                "operationId": "storeRole",
                "requestBody": {
                    "description": "Role to create",
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "data": {
                                        "$ref": "#/components/schemas/Role"
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Role created",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "422": {
                        "description": "Validation exception"
                    }
                },
                "security": [
                    {
                        "bearerAuth": []
                    }
                ]
            }
        },
        "/roles/{role_id}": {
            "get": {
                "tags": [
                    "Roles"
                ],
                "summary": "Show a Role from his Id",
                "description": "Show a Role from his Id",
                "operationId": "showRole",
                "parameters": [
                    {
                        "$ref": "#/components/parameters/Role--id"
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Successful operation",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "data": {
                                            "$ref": "#/components/schemas/Role"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    },
                    "404": {
                        "description": "Role not found"
                    }
                }
            }
        },
...
    },
    "components": {
        "schemas": {
...
            "Product": {
                "title": "Product",
                "description": "Product model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Product",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name": {
                        "title": "name",
                        "description": "name of Product",
                        "type": "string",
                        "example": "Macbook Pro"
                    },
                    "sku": {
                        "title": "sku",
                        "description": "sku of Product",
                        "type": "string",
                        "example": "MCBPRO2022"
                    },
                    "price": {
                        "title": "price",
                        "description": "price of Product",
                        "type": "integer",
                        "example": "99"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
...
            "Role": {
                "title": "Roles",
                "description": "Role model",
                "required": [],
                "properties": {
                    "id": {
                        "title": "id",
                        "description": "id of Role",
                        "type": "integer",
                        "readOnly": "true",
                        "example": "1"
                    },
                    "name_role": {
                        "title": "name_role",
                        "description": "name of Role",
                        "type": "string",
                        "example": "Role user"
                    },
                    "code_role": {
                        "title": "code_role",
                        "description": "Code of role",
                        "type": "integer",
                        "example": "0"
                    },
                    "created_at": {
                        "title": "created_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    },
                    "updated_at": {
                        "title": "updated_at",
                        "type": "dateTime",
                        "readOnly": "true",
                        "example": "2022-07-04T02:41:42.336Z"
                    }
                },
                "type": "object"
            },
            "Roles": {
                "title": "Roles",
                "properties": {
                    "data": {
                        "title": "data",
                        "type": "array",
                        "items": {
                            "$ref": "#/components/schemas/Role"
                        }
                    }
                },
                "type": "object"
            },
...
        },
        "parameters": {
...
            "Product--id": {
                "name": "Product_id",
                "in": "path",
                "description": "Id of Product",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            },
            "Role--id": {
                "name": "role_id",
                "in": "path",
                "description": "Id of Role",
                "required": true,
                "schema": {
                    "type": "integer",
                    "example": "1"
                }
            }
        },
        "securitySchemes": {
            "bearerAuth": {
                "type": "http",
                "name": "bearerAuth",
                "in": "header",
                "bearerFormat": "JWT",
                "scheme": "bearer"
            }
        }
    }
}